+++ kernel/stdlib/src/stdio.c
@@ -249,7 +249,7 @@ void perror(const char* format, ...) {
  * \returns the number of read characters (excluding the null terminate). Else
  * return -1.
  */
-int64_t getline(char** str, size_t* size, int* stream) {
+int64_t getline(char** str, size_t* size, uint64_t* stream) {
   // Check for NULL pointers from params
   if (str == NULL || size == NULL || stream == NULL) {
     return -1;
